+++ src/Score.php
@@ -40,35 +40,57 @@ class Score
     /**
      * Reference to the test the score belongs to.
      *
-     * @var Test
+     * @var TestCase
      */
-    private Test $test;
+    private TestCase $test;
+    
+    /**
+     * Array containing score information.
+     *
+     * @var array
+     */
+    private array $results = [];
     
     /**
      * Score constructor.
      *
-     * @param Test $test The Test this score belongs to.
+     * @param  TestCase  $test  The Test this score belongs to.
      */
-    public function __construct(Test $test)
+    public function __construct(TestCase $test)
     {
         $this->test = $test;
         
-        /**
-         * Set the default score to the minimal score.
-         */
+        $this->reset();
+    }
+    
+    /**
+     * Reset the values to default.
+     *
+     * @return void
+     */
+    public function reset()
+    {
         $this->score = $this->minScore();
+        $this->increments = 0;
+        $this->results = [];
     }
     
     /**
      * Increment the score by $score amount.
      *
-     * @param mixed $value The value to increment the score with (float/double or int).
+     * @param  mixed   $value       The value to increment the score with (float/double or int).
+     * @param  string  $motivation  A motivation for this score.
      *
      * @return void
      */
-    public function increment($value)
+    public function increment($value, $motivation = '')
     {
         $this->score += $value;
+        $this->results[$this->increments] = [
+          'score' => $value,
+          'increment' => $this->increments,
+          'motivation' => $motivation,
+        ];
         $this->increments++;
     }
     
@@ -106,6 +128,16 @@ class Score
     }
     
     /**
+     * Return information about the scoreszz.
+     *
+     * @return array
+     */
+    public function getScoreInfo(): array
+    {
+        return $this->results;
+    }
+    
+    /**
      * Return the minimal score.
      *
      * @return mixed
@@ -136,9 +168,19 @@ class Score
     }
     
     /**
+     * Return the parent class instance.
+     *
+     * @return TestCase
+     */
+    public function getTest(): TestCase
+    {
+        return $this->test;
+    }
+    
+    /**
      * Set to score to a given value.
      *
-     * @param int $value The value to set as the score.
+     * @param  int  $value  The value to set as the score.
      *
      * @return void
      */